requirements.txt
alembic==1.14.1
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dominate==2.9.1
Flask==3.1.0
Flask-Bootstrap4==4.0.2
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Moment==1.0.6
flask-paginate==2024.4.12
Flask-Session==0.8.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
msgspec==0.19.0
packaging==24.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
simple-websocket==1.1.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
visitor==0.1.3
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
